feat: refactor out components
